Put Exception API surface in place, per #36. Next PR will get into the details of the error codes and hanlding exceptions across the source.
