+++ src/Microsoft.Identity.Client/MsalException.cs
@@ -26,7 +26,6 @@
 //------------------------------------------------------------------------------
 
 using System;
-using System.Collections.Generic;
 using System.Globalization;
 using Microsoft.Identity.Client.Internal;
 
@@ -37,41 +36,11 @@ namespace Microsoft.Identity.Client
     /// </summary>
     public class MsalException : Exception
     {
-        private static readonly Dictionary<string, string> ErrorMessages = new Dictionary<string, string>
-        {
-            [MsalError.InvalidCredentialType] = MsalErrorMessage.InvalidCredentialType,
-            [MsalError.IdentityProtocolLoginUrlNull] = MsalErrorMessage.IdentityProtocolLoginUrlNull,
-            [MsalError.IdentityProtocolMismatch] = MsalErrorMessage.IdentityProtocolMismatch,
-            [MsalError.EmailAddressSuffixMismatch] = MsalErrorMessage.EmailAddressSuffixMismatch,
-            [MsalError.IdentityProviderRequestFailed] = MsalErrorMessage.IdentityProviderRequestFailed,
-            [MsalError.StsTokenRequestFailed] = MsalErrorMessage.StsTokenRequestFailed,
-            [MsalError.EncodedTokenTooLong] = MsalErrorMessage.EncodedTokenTooLong,
-            [MsalError.StsMetadataRequestFailed] = MsalErrorMessage.StsMetadataRequestFailed,
-            [MsalError.AuthorityNotInValidList] = MsalErrorMessage.AuthorityNotInValidList,
-            [MsalError.UnsupportedUserType] = MsalErrorMessage.UnsupportedUserType,
-            [MsalError.UnknownUser] = MsalErrorMessage.UnknownUser,
-            [MsalError.UserRealmDiscoveryFailed] = MsalErrorMessage.UserRealmDiscoveryFailed,
-            [MsalError.AccessingWsMetadataExchangeFailed] = MsalErrorMessage.AccessingMetadataDocumentFailed,
-            [MsalError.ParsingWsMetadataExchangeFailed] = MsalErrorMessage.ParsingMetadataDocumentFailed,
-            [MsalError.WsTrustEndpointNotFoundInMetadataDocument] = MsalErrorMessage.WsTrustEndpointNotFoundInMetadataDocument,
-            [MsalError.ParsingWsTrustResponseFailed] = MsalErrorMessage.ParsingWsTrustResponseFailed,
-            [MsalError.AuthenticationCanceled] = MsalErrorMessage.AuthenticationCanceled,
-            [MsalError.NetworkNotAvailable] = MsalErrorMessage.NetworkIsNotAvailable,
-            [MsalError.AuthenticationUiFailed] = MsalErrorMessage.AuthenticationUiFailed,
-            [MsalError.UserInteractionRequired] = MsalErrorMessage.UserInteractionRequired,
-            [MsalError.MissingFederationMetadataUrl] = MsalErrorMessage.MissingFederationMetadataUrl,
-            [MsalError.IntegratedAuthFailed] = MsalErrorMessage.IntegratedAuthFailed,
-            [MsalError.UnauthorizedResponseExpected] = MsalErrorMessage.UnauthorizedResponseExpected,
-            [MsalError.MultipleTokensMatched] = MsalErrorMessage.MultipleTokensMatched,
-            [MsalError.PasswordRequiredForManagedUserError] = MsalErrorMessage.PasswordRequiredForManagedUserError,
-            [MsalError.GetUserNameFailed] = MsalErrorMessage.GetUserNameFailed,
-            // MsalErrorMessage.Unknown will be set as the default error message in GetErrorMessage(string errorCode).
-        };
 
         /// <summary>
         /// Initializes a new instance of the exception class.
         /// </summary>
-        public MsalException()
+        internal MsalException()
             : base(MsalErrorMessage.Unknown)
         {
             ErrorCode = MsalError.Unknown;
@@ -85,8 +54,7 @@ namespace Microsoft.Identity.Client
         /// The error code returned by the service or generated by client. This is the code you can rely on
         /// for exception handling.
         /// </param>
-        public MsalException(string errorCode)
-            : base(GetErrorMessage(errorCode))
+        internal MsalException(string errorCode)
         {
             ErrorCode = errorCode;
         }
@@ -100,7 +68,7 @@ namespace Microsoft.Identity.Client
         /// for exception handling.
         /// </param>
         /// <param name="message">The error message that explains the reason for the exception.</param>
-        public MsalException(string errorCode, string message)
+        internal MsalException(string errorCode, string message)
             : base(message)
         {
             ErrorCode = errorCode;
@@ -115,32 +83,13 @@ namespace Microsoft.Identity.Client
         /// The error code returned by the service or generated by client. This is the code you can rely on
         /// for exception handling.
         /// </param>
+        /// <param name="errorMessage">The error message that explains the reason for the exception.</param>
         /// <param name="innerException">
         /// The exception that is the cause of the current exception, or a null reference if no inner
         /// exception is specified. It may especially contain the actual error message returned by the service.
         /// </param>
-        public MsalException(string errorCode, Exception innerException)
-            : base(GetErrorMessage(errorCode), innerException)
-        {
-            ErrorCode = errorCode;
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the exception class with a specified
-        /// error code, error message and a reference to the inner exception that is the cause of
-        /// this exception.
-        /// </summary>
-        /// <param name="errorCode">
-        /// The error code returned by the service or generated by client. This is the code you can rely on
-        /// for exception handling.
-        /// </param>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        /// <param name="innerException">
-        /// The exception that is the cause of the current exception, or a null reference if no inner
-        /// exception is specified. It may especially contain the actual error message returned by the service.
-        /// </param>
-        public MsalException(string errorCode, string message, Exception innerException)
-            : base(message, innerException)
+        internal MsalException(string errorCode, string errorMessage, Exception innerException)
+            : base(errorMessage, innerException)
         {
             ErrorCode = errorCode;
         }
@@ -158,21 +107,6 @@ namespace Microsoft.Identity.Client
         public override string ToString()
         {
             return base.ToString() + string.Format(CultureInfo.InvariantCulture, "\n\tErrorCode: {0}", ErrorCode);
-        }
-
-        /// <summary>
-        /// Gets the Error Message
-        /// </summary>
-        protected static string GetErrorMessage(string errorCode)
-        {
-            string message = ErrorMessages.ContainsKey(errorCode) ? ErrorMessages[errorCode] : MsalErrorMessage.Unknown;
-            return String.Format(CultureInfo.InvariantCulture, "{0}: {1}", errorCode, message);
-        }
-
-        internal enum ErrorFormat
-        {
-            Json,
-            Other
         }
     }
 }
