+++ src/Microsoft.Identity.Client/MsalServiceException.cs
@@ -27,9 +27,6 @@
 
 using System;
 using System.Globalization;
-using System.Net;
-using System.Threading.Tasks;
-using Microsoft.Identity.Client.Internal.Interfaces;
 
 namespace Microsoft.Identity.Client
 {
@@ -38,38 +35,6 @@ namespace Microsoft.Identity.Client
     /// </summary>
     public class MsalServiceException : MsalException
     {
-        /// <summary>
-        /// Initializes a new instance of the exception class with a specified
-        /// error code and error message.
-        /// </summary>
-        /// <param name="errorCode">
-        /// The protocol error code returned by the service or generated by client. This is the code you
-        /// can rely on for exception handling.
-        /// </param>
-        /// <param name="message">The error message that explains the reason for the exception.</param>
-        public MsalServiceException(string errorCode, string message)
-            : base(errorCode, message)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the exception class with a specified
-        /// error code and a reference to the inner exception that is the cause of
-        /// this exception.
-        /// </summary>
-        /// <param name="errorCode">
-        /// The protocol error code returned by the service or generated by client. This is the code you
-        /// can rely on for exception handling.
-        /// </param>
-        /// <param name="innerException">
-        /// The exception that is the cause of the current exception, or a null reference if no inner
-        /// exception is specified. It may especially contain the actual error message returned by the service.
-        /// </param>
-        internal MsalServiceException(string errorCode, Exception innerException)
-            : this(errorCode, GetErrorMessage(errorCode), innerException)
-        {
-        }
-
         /// <summary>
         /// Initializes a new instance of the exception class with a specified
         /// error code, error message and a reference to the inner exception that is the cause of
@@ -92,18 +57,62 @@ namespace Microsoft.Identity.Client
         }
 
         /// <summary>
+        /// Initializes a new instance of the exception class with a specified
+        /// error code, error message and a reference to the inner exception that is the cause of
+        /// this exception.
+        /// </summary>
+        /// <param name="errorCode">
+        /// The protocol error code returned by the service or generated by client. This is the code you
+        /// can rely on for exception handling.
+        /// </param>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="innerException">
+        /// The exception that is the cause of the current exception, or a null reference if no inner
+        /// exception is specified. It may especially contain the actual error message returned by the service.
+        /// </param>
+        internal MsalServiceException(string errorCode, string message, int statusCode,
+            Exception innerException)
+            : base(
+                errorCode, message, innerException)
+        {
+            StatusCode = statusCode;
+        }
+
+
+        /// <summary>
+        /// Initializes a new instance of the exception class with a specified
+        /// error code, error message and a reference to the inner exception that is the cause of
+        /// this exception.
+        /// </summary>
+        /// <param name="errorCode">
+        /// The protocol error code returned by the service or generated by client. This is the code you
+        /// can rely on for exception handling.
+        /// </param>
+        /// <param name="message">The error message that explains the reason for the exception.</param>
+        /// <param name="statusCode">The status code of the request.</param>
+        /// <param name="claims">The claims challenge returned back from the service.</param>
+        /// <param name="innerException">
+        /// The exception that is the cause of the current exception, or a null reference if no inner
+        /// exception is specified. It may especially contain the actual error message returned by the service.
+        /// </param>
+        internal MsalServiceException(string errorCode, string message, int statusCode, string claims,
+            Exception innerException)
+            : this(
+                errorCode, message, statusCode, innerException)
+        {
+            Claims = claims;
+        }
+
+        /// <summary>
         /// Gets the status code returned from http layer. This status code is either the HttpStatusCode in the inner
         /// HttpRequestException response or
         /// NavigateError Event Status Code in browser based flow (See
         /// http://msdn.microsoft.com/en-us/library/bb268233(v=vs.85).aspx).
         /// You can use this code for purposes such as implementing retry logic or error investigation.
         /// </summary>
-        public int StatusCode { get; set; }
+        public int StatusCode { get; } = 0;
 
-        /// <summary>
-        /// Gets the specific error codes that may be returned by the service.
-        /// </summary>
-        public string[] ServiceErrorCodes { get; set; }
+        public string Claims { get; }
 
         /// <summary>
         /// Creates and returns a string representation of the current exception.
@@ -111,7 +120,7 @@ namespace Microsoft.Identity.Client
         /// <returns>A string representation of the current exception.</returns>
         public override string ToString()
         {
-            return base.ToString() + string.Format(CultureInfo.InvariantCulture, "\n\tStatusCode: {0}", StatusCode);
+            return base.ToString() + string.Format(CultureInfo.InvariantCulture, "\n\tStatusCode: {0}\n\tClaims: {1}", StatusCode, Claims);
         }
     }
 }
